move to static LineBoundsHelper class.
